requirements.txt
feedparser
google-generativeai
beautifulsoup4
